requirements.txt
# Core ML Libraries
ultralytics==8.0.196
torch>=2.0.0
torchvision>=0.15.0
opencv-python==4.8.1.78
numpy==1.24.3
pillow==11.1.0
pdf2image==1.16.3
pillow-heif>=0.15.0
PyMuPDF>=1.23.0
easyocr==1.7.1
requests>=2.31.0
google-generativeai>=0.8.0

# Backend
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0

# Utilities
pyyaml==6.0.1
tqdm==4.66.1
pandas==2.1.1
matplotlib==3.8.0
seaborn==0.13.0

# Optional: для работы с разными форматами
scikit-learn==1.3.1
